+++ pkg/firewall/firewall_test.go
@@ -1,0 +1,9 @@
+package firewall_test
+
+import (
+	. "github.com/onsi/ginkgo"
+)
+
+var _ = Describe("Firewall", func() {
+
+})
