+++ pkg/firewall/iptables.go
@@ -1,0 +1,2 @@
+// Package firewall handles the firewall and forwarding configuration
+package firewall
