--- pkg/iptables/datatypes.go
@@ -1,2 +1,0 @@
-// Package iptables handles the firewall and forwarding configuration
-package iptables
