--- pkg/iptables/iptables_test.go
@@ -1,9 +1,0 @@
-package iptables_test
-
-import (
-	. "github.com/onsi/ginkgo"
-)
-
-var _ = Describe("Iptables", func() {
-
-})
